work on float multiply

+++ assignment2/functions.h
@@ -1,4 +1,4 @@
 
 void print_bits(int x);
-long int_multiply(long x, long y);
+long int_multiply(int x, int y);
 float float_multiply(float a, float b);
